refactor(api): getting lost on llm package abstraction, need to put this on paper
