--- api/internal/llm/config.go
@@ -1,13 +1,0 @@
-package llm
-
-type LLMConfig struct {
-	model      string
-	isStreamed bool
-}
-
-func NewLLMConfig() LLMConfig {
-	return LLMConfig{
-		model:      model,
-		isStreamed: isStreamed,
-	}
-}
