--- api/internal/llm/llm.go
@@ -1,14 +1,0 @@
-package llm
-
-// const (
-// 	baseUrl    = "http://localhost:11434"
-// 	model      = "mistral"
-// 	isStreamed = false
-// )
-
-type ModelType string
-type BaseUrlType string
-
-const (
-	ollamaBaseUrl BaseUrlType = "http://localhost:11434"
-)
